# Populate names of colliding entities in contact points message