--- a/src/systems/physics/Physics.cc
+++ b/src/systems/physics/Physics.cc
@@ -3259,7 +3259,11 @@ void PhysicsPrivate::UpdateCollisions(EntityComponentManager &_ecm)
         {
           msgs::Contact *contactMsg = contactsComp.add_contact();
           contactMsg->mutable_collision1()->set_id(_collEntity1);
+	  auto _collEntity1Name = _ecm.Component<components::Name>(_collEntity1)->Data();
+          contactMsg->mutable_collision1()->set_name(_collEntity1Name);
           contactMsg->mutable_collision2()->set_id(collEntity2);
+	  auto collEntity2Name = _ecm.Component<components::Name>(collEntity2)->Data();
+          contactMsg->mutable_collision2()->set_name(collEntity2Name);
           for (const auto &contact : contactData)
           {
             auto *position = contactMsg->add_position();