--- a/src/systems/physics/Physics.cc
+++ b/src/systems/physics/Physics.cc
@@ -3259,11 +3259,11 @@ void PhysicsPrivate::UpdateCollisions(EntityComponentManager &_ecm)
         {
           msgs::Contact *contactMsg = contactsComp.add_contact();
           contactMsg->mutable_collision1()->set_id(_collEntity1);
-	  auto _collEntity1Name = _ecm.Component<components::Name>(_collEntity1)->Data();
-          contactMsg->mutable_collision1()->set_name(_collEntity1Name);
+          contactMsg->mutable_collision1()->set_name(
+	    scopedName(_collEntity1, _ecm, ":", 1));
           contactMsg->mutable_collision2()->set_id(collEntity2);
-	  auto collEntity2Name = _ecm.Component<components::Name>(collEntity2)->Data();
-          contactMsg->mutable_collision2()->set_name(collEntity2Name);
+          contactMsg->mutable_collision2()->set_name(
+	    scopedName(collEntity2, _ecm, ":", 1));
           for (const auto &contact : contactData)
           {
             auto *position = contactMsg->add_position();