@peci1 2022-02-23
Thanks for the PR, it looks useful! However, I would really like to see a performance analysis of this. Try to create a world with say 10 000 contacts and measure the update rate with and without this PR...