@adityapande-1995 2022-02-24
## Performance benchmarking : 

### Generate world file 
I used the following ``erb`` template to generate a test world file : 

```
<?xml version="1.0" ?>
<!--
  To generate sdf file using this erb template, run `erb -T 1 shapes_population.sdf.erb > shapes_population.sdf`
-->
<sdf version="1.6">
  <world name="shapes">
    <physics name="1ms" type="ignored">
      <max_step_size>0.001</max_step_size>
      <real_time_factor>1.0</real_time_factor>
    </physics>
    <plugin
      filename="ignition-gazebo-physics-system"
      name="ignition::gazebo::systems::Physics">
    </plugin>
    <plugin
      filename="ignition-gazebo-sensors-system"
      name="ignition::gazebo::systems::Sensors">
      <render_engine>ogre2</render_engine>
    </plugin>
    <plugin
      filename="ignition-gazebo-scene-broadcaster-system"
      name="ignition::gazebo::systems::SceneBroadcaster">
    </plugin>
    <plugin
      filename="ignition-gazebo-contact-system"
      name="ignition::gazebo::systems::Contact">
    </plugin>

    <scene>
      <ambient>1.0 1.0 1.0</ambient>
      <background>0.8 0.8 0.8</background>
    </scene>

    <light type="directional" name="sun">
      <cast_shadows>true</cast_shadows>
      <pose>0 0 10 0 0 0</pose>
      <diffuse>0.8 0.8 0.8 1</diffuse>
      <specular>0.8 0.8 0.8 1</specular>
      <attenuation>
        <range>1000</range>
        <constant>0.9</constant>
        <linear>0.01</linear>
        <quadratic>0.001</quadratic>
      </attenuation>
      <direction>-0.5 0.1 -0.9</direction>
    </light>

    <model name="ground_plane">
      <static>true</static>
      <link name="link">
        <collision name="collision">
          <geometry>
            <plane>
              <normal>0 0 1</normal>
              <size>100 100</size>
            </plane>
          </geometry>
        </collision>
        <visual name="visual">
          <geometry>
            <plane>
              <normal>0 0 1</normal>
              <size>100 100</size>
            </plane>
          </geometry>
          <material>
            <ambient>0.8 0.8 0.8 1</ambient>
            <diffuse>0.8 0.8 0.8 1</diffuse>
            <specular>0.8 0.8 0.8 1</specular>
          </material>
        </visual>
      </link>
    </model>

    <%
      # number of population
      n = 100
      for m in (0..n)
        geom_1 = 'box'
    %>

    <model name="<%= geom_1 %>_<%= m.to_s %>">
      <pose><%= 0 + m*1.5 %> 0 0.5 0 0 0</pose>
      <link name="box_link">
        <inertial>
          <inertia>
            <ixx>1</ixx>
            <ixy>0</ixy>
            <ixz>0</ixz>
            <iyy>1</iyy>
            <iyz>0</iyz>
            <izz>1</izz>
          </inertia>
          <mass>1.0</mass>
        </inertial>
        <collision name="box_collision">
          <geometry>
            <box>
              <size>1 1 1</size>
            </box>
          </geometry>
        </collision>
        
        <sensor name='sensor_contact_<%= m.to_s %>' type='contact'>
	<contact>
	  <collision>box_collision</collision>
	  <topic>/robot_collision</topic>
        </contact>
        <always_on>1</always_on>
        <update_rate>1000</update_rate>
       </sensor>

        <visual name="box_visual">
          <geometry>
            <box>
              <size>1 1 1</size>
            </box>
          </geometry>
          <material>
            <ambient>1 0 0 1</ambient>
            <diffuse>1 0 0 1</diffuse>
            <specular>1 0 0 1</specular>
          </material>
        </visual>
      </link>
    </model>

    <%
      end
    %>
  </world>
</sdf>
```

Then, ``erb -T 1 template.erb > test.sdf ; ign gazebo -r -s test.sdf`` runs the simulation.

### Data collection

I collected the real time factor data in a text file (let the simulation run for around 30 seconds) : ``ign topic -e -t /stats | grep real_time_factor | cut -c 19- > data.txt``, and got an average of that data using : ``awk '{ sum += $1; n++ } END { if (n > 0) print sum / n; print n}' data.txt ``

### Results

| Trails | Avg real time factor ``gazebo6`` branch | Avg real time factor current branch | % change |
| ------------- | ------------- | ------------- |
| 1. |  0.0918426 | 0.0894021 | -2.65 |
| 2.  | 0.0914473 |  0.0910263 | -0.46 |
| 3.  | 0.0927287 |  0.0912242 | -1.62 |
| 4.  | 0.0926891|  0.0915801 ​|  -1.19 |