@adityapande-1995 2022-02-24
Added an optional parameter so that the names of colliding entities will not be populated and performance won't be affected. This parameter remains true by default.