@azeey 2022-02-25
Also could you document the parameter in `Physics.hh`?