@adityapande-1995 2022-02-25
I've added the docstring, not sure if that is the preferred style though.