@peci1 2022-02-26
> Performance benchmarking :

Thanks a lot for the benchmark, it's good to see the performance isn't affected that heavily. In normal situations with just a few contacts, the difference would be negligible.